feat: implement JDBC persistence layer and domain refactoring

+++ src/main/java/controller/ProjectController.java
@@ -2,7 +2,6 @@ package controller;
 
 import model.Project;
 import service.interfaces.IProjectService;
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -33,7 +32,8 @@ public class ProjectController {
         service.delete(id);
     }
 
-    public BigDecimal calculateHrCost(Integer projectId) {
+     // Task 1: Cost Calculation
+    public Double calculateHrCost(Integer projectId) {
         return service.calculateProjectHRCost(projectId);
     }
 }
