+++ src/main/java/service/interfaces/IProjectService.java
@@ -2,10 +2,10 @@ package service.interfaces;
 
 import model.Project;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 public interface IProjectService extends IService<Project> {
-    BigDecimal calculateProjectHRCost(int projectId);
+    Double calculateProjectHRCost(int projectId);
+
     List<Project> getProjectsByDepartment(int departmentId, String sortBy);
 }
